fix(schema): repair 17 dangling cross-schema calls; pg_regress 1/20 -> 20/20
pg_regress was passing exactly one test of twenty. None of it was stale
expectations -- every failure traced to a call the purpose-schema move left
pointing at a name that no longer exists there.

  taxonomy.taxonomy_tree      -> taxonomy.tree (renamed on the move)
  consensus.relation_band_types -> laplace.relation_band_types  (x3)
  consensus.{attestation_response, attestation_response_type,
             attestation_unary_response, attestation_unary_response_type,
             consensus_out_labeled, consensus_partition_pressure,
             entity_facets, entity_physicalities,
             entity_physicality_coords}                        -> ops.*
  consensus.laplace_entity_attestations -> ops.entity_attestations
  lexical.laplace_translations          -> lexical.translations
  structural.laplace_entities_at_depth  -> structural.entities_at_depth
  taxonomy.laplace_ancestry             -> taxonomy.ancestry
  scoped_consensus:30  unqualified consensus_fold + untyped NULLs
  seed_relation_partitions  unqualified relation_type_id at install time

taxonomy_tree was not only a test break: NpgsqlSubstrateReads.cs:52 issues
FROM taxonomy.taxonomy_tree(@id), so the C# read path raised 42883 at runtime.

WHY ONE NAME COST SO MUCH. pg_regress runs each file in a transaction, so the
first dangling call aborts every statement after it -- ops_reads lost eight
assertions to a single rename, and the file still "ran". A rename that nothing
calls looks identical to a rename whose callers all moved, right up until the
transaction aborts.

The expected outputs are REGENERATED FROM THE HARNESS RUN, not hand-edited to
match. Before copying, every file was diffed to separate echoed SQL from result
rows: the only non-echo change in the whole suite is structural_surface's
f -> t, which is the test correctly finding cluster() now that it lives in
structural rather than laplace. Hand-substituting these would have enshrined an
ERROR plus a cascade of skipped assertions as the expectation, and it would have
reported green.

Verified: build, install, CREATE EXTENSION on a fresh database, ISA gate 0,
documentation governance ok, pg_regress 20/20, and every cross-schema call in
the extension resolves against pg_proc.

+++ extension/laplace_substrate/tests/sql/consensus_signed.sql
@@ -56,13 +56,13 @@ BEGIN
     -- spine uses: consensus_fold(is_seed, rating, rd, volatility, phi, games,
     -- sum_score, tau). Trust enters as the opponent RD (phi); no seed row =
     -- fold starts from the neutral prior, exactly as a fresh consensus cell.
-    SELECT (laplace.consensus_fold(false, NULL, NULL, NULL, phi_trust, 1, s_conf,     NULL)).rating INTO mu_conf;
-    SELECT (laplace.consensus_fold(false, NULL, NULL, NULL, phi_trust, 1, s_ref,      NULL)).rating INTO mu_ref;
-    SELECT (laplace.consensus_fold(false, NULL, NULL, NULL, phi_trust, 1, s_draw,     NULL)).rating INTO mu_draw;
-    SELECT (laplace.consensus_fold(false, NULL, NULL, NULL, phi_trust, 1, s_conf,     NULL)).rating INTO mu_trust;
-    SELECT (laplace.consensus_fold(false, NULL, NULL, NULL, phi_crank, 1, s_conf,     NULL)).rating INTO mu_crank;
-    SELECT (laplace.consensus_fold(false, NULL, NULL, NULL, phi_trust, 8, s_conf * 8, NULL)).rating INTO mu_many;
-    SELECT (laplace.consensus_fold(false, NULL, NULL, NULL, phi_trust, 1, s_conf,     NULL)).rating INTO mu_one;
+    SELECT (laplace.consensus_fold(false, NULL::bigint, NULL::bigint, NULL::bigint, phi_trust, 1, s_conf, NULL::bigint)).rating INTO mu_conf;
+    SELECT (laplace.consensus_fold(false, NULL::bigint, NULL::bigint, NULL::bigint, phi_trust, 1, s_ref, NULL::bigint)).rating INTO mu_ref;
+    SELECT (laplace.consensus_fold(false, NULL::bigint, NULL::bigint, NULL::bigint, phi_trust, 1, s_draw, NULL::bigint)).rating INTO mu_draw;
+    SELECT (laplace.consensus_fold(false, NULL::bigint, NULL::bigint, NULL::bigint, phi_trust, 1, s_conf, NULL::bigint)).rating INTO mu_trust;
+    SELECT (laplace.consensus_fold(false, NULL::bigint, NULL::bigint, NULL::bigint, phi_crank, 1, s_conf, NULL::bigint)).rating INTO mu_crank;
+    SELECT (laplace.consensus_fold(false, NULL::bigint, NULL::bigint, NULL::bigint, phi_trust, 8, s_conf * 8, NULL::bigint)).rating INTO mu_many;
+    SELECT (laplace.consensus_fold(false, NULL::bigint, NULL::bigint, NULL::bigint, phi_trust, 1, s_conf, NULL::bigint)).rating INTO mu_one;
 
     RAISE NOTICE 'confirm μ=% refute μ=% draw μ=% (neutral=%)', mu_conf, mu_ref, mu_draw, neutral;
     RAISE NOTICE 'trusted μ=% crank μ=% | 8-games μ=% 1-game μ=%', mu_trust, mu_crank, mu_many, mu_one;
